# services/meta-center/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 元数据服务API
  description: |
    元数据驱动的低代码平台核心服务，提供统一的元数据管理、动态数据操作、业务规则引擎、权限控制等能力。
    
    ## 核心功能
    - **元数据管理**: 支持元数据对象、属性、关系、规则的全生命周期管理
    - **动态数据服务**: 基于元数据的动态CRUD操作，支持表单渲染和数据校验
    - **业务规则引擎**: 动态配置和执行业务规则，支持多种表达式类型
    - **权限与安全**: 细粒度权限控制和数据脱敏
    - **多租户支持**: 租户级别的数据和配置隔离
    
    ## 版本与兼容性
    - API版本遵循语义化版本规范
    - 向后兼容性保证，废弃接口将提前通知
    - 支持接口版本协商和灰度发布
  version: 1.0.0
  contact:
    name: 元数据服务团队
    email: metadata-team@aixone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.aixone.com/meta-center/v1
    description: 生产环境
  - url: https://api-staging.aixone.com/meta-center/v1
    description: 测试环境
  - url: http://localhost:8080/meta-center/v1
    description: 本地开发环境
security:
  - BearerAuth: []
  - TenantHeader: []
paths:
...
  /metadata/objects/{objectId}:
    get:
      tags:
        - 元数据管理
      summary: 获取元数据对象详情
      description: 获取指定元数据对象的完整信息，包括属性、关系、规则等
      parameters:
        - $ref: '#/components/parameters/ObjectIdParam'
        - name: includeRelations
          in: query
          description: 是否包含关系信息
          schema:
            type: boolean
            default: true
        - name: includeRules
          in: query
          description: 是否包含规则信息
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaObjectDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - 元数据管理
      summary: 更新元数据对象
      description: 更新元数据对象的完整信息，支持增量更新
      parameters:
        - $ref: '#/components/parameters/ObjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMetaObjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaObjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    
    delete:
      tags:
        - 元数据管理
      summary: 删除元数据对象
      description: 删除指定的元数据对象，支持软删除和硬删除
      parameters:
        - $ref: '#/components/parameters/ObjectIdParam'
        - name: hardDelete
          in: query
          description: 是否硬删除
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: 删除成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 存在依赖关系，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
...
  /metadata/objects/{objectId}/attributes/{attributeName}:
    put:
      tags:
        - 元数据管理
      summary: 更新对象属性
      parameters:
        - $ref: '#/components/parameters/ObjectIdParam'
        - name: attributeName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMetaAttributeRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAttribute'
    
    delete:
      tags:
        - 元数据管理
      summary: 删除对象属性
      parameters:
        - $ref: '#/components/parameters/ObjectIdParam'
        - name: attributeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证token
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: 租户标识
  parameters:
...
    ObjectIdParam:
      name: objectId
      in: path
      required: true
      description: 元数据对象ID
      schema:
        type: string
  schemas:
...
    MetadataType:
      type: string
      enum: [business, technical, administrative, reference]
      description: |
        元数据主类型：
        - business: 业务元数据
        - technical: 技术元数据  
        - administrative: 管理元数据
        - reference: 参考元数据
    LifecycleStatus:
      type: string
      enum: [draft, published, archived, deprecated]
      description: |
        生命周期状态：
        - draft: 草稿
        - published: 已发布
        - archived: 已归档
        - deprecated: 已废弃

    # ==================== 元数据对象模型 ====================
    MetaObject:
      type: object
      required: [id, name, objectType, type]
      properties:
        id:
          type: string
          description: 唯一标识
        name:
          type: string
          description: 对象名称
        objectType:
          type: string
          enum: [Entity, Attribute, Relation, Rule, Dictionary]
          description: 对象类型
        type:
          $ref: '#/components/schemas/MetadataType'
        subType:
          type: string
          description: 子类型
        description:
          type: string
          description: 描述信息
        tags:
          type: array
          items:
            type: string
          description: 标签列表
        lifecycle:
          $ref: '#/components/schemas/LifecycleStatus'
        status:
          type: string
          enum: [enabled, disabled, locked]
          description: 当前状态
        owner:
          type: string
          description: 负责人
        responsible:
          type: string
          description: 责任人
        qualityScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 质量分数
        complianceLevel:
          type: string
          enum: [A, B, C, D]
          description: 合规等级
        version:
          type: integer
          minimum: 1
          description: 版本号
        tenantId:
          type: string
          description: 租户ID
        createdBy:
          type: string
          description: 创建人
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedBy:
          type: string
          description: 更新人
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        extensionFields:
          type: object
          additionalProperties: true
          description: 扩展字段
    MetaAttribute:
      type: object
      required: [name, label, schema]
      properties:
        name:
          type: string
          description: 属性名
        label:
          type: string
          description: 显示名
        schema:
          type: object
          description: JSON Schema定义
          properties:
            type:
              type: string
              enum: [string, number, integer, boolean, array, object]
            format:
              type: string
              description: 格式约束
            pattern:
              type: string
              description: 正则表达式
            minimum:
              type: number
              description: 最小值
            maximum:
              type: number
              description: 最大值
            minLength:
              type: integer
              description: 最小长度
            maxLength:
              type: integer
              description: 最大长度
            enum:
              type: array
              description: 枚举值
        required:
          type: boolean
          default: false
          description: 是否必填
        defaultValue:
          description: 默认值
        enumRef:
          type: string
          description: 引用的字典/枚举ID
        validationRules:
          type: array
          items:
            $ref: '#/components/schemas/ValidationRule'
          description: 校验规则
        ui:
          $ref: '#/components/schemas/UIConfig'
          description: UI配置
        security:
          $ref: '#/components/schemas/SecurityConfig'
          description: 安全配置
    ValidationRule:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [regex, range, length, custom, unique]
          description: 校验类型
        pattern:
          type: string
          description: 正则表达式（regex类型）
        min:
          type: number
          description: 最小值（range类型）
        max:
          type: number
          description: 最大值（range类型）
        minLength:
          type: integer
          description: 最小长度（length类型）
        maxLength:
          type: integer
          description: 最大长度（length类型）
        expression:
          type: string
          description: 自定义表达式（custom类型）
        message:
          type: string
          description: 错误提示信息
        severity:
          type: string
          enum: [error, warning, info]
          default: error
          description: 严重程度
    UIConfig:
      type: object
      properties:
        inputType:
          type: string
          enum: [text, password, email, number, tel, url, textarea, select, checkbox, radio, date, datetime, file, switch]
          description: 输入控件类型
        placeholder:
          type: string
          description: 占位符
        order:
          type: integer
          description: 显示顺序
        visible:
          type: boolean
          default: true
          description: 是否可见
        readonly:
          type: boolean
          default: false
          description: 是否只读
        disabled:
          type: boolean
          default: false
          description: 是否禁用
        width:
          type: string
          description: 宽度设置
        options:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              value:
                description: 值
          description: 选项列表（select/radio类型）
        style:
          type: object
          additionalProperties: true
          description: 自定义样式
    SecurityConfig:
      type: object
      properties:
        masking:
          type: boolean
          default: false
          description: 是否需要脱敏
        maskingType:
          type: string
          enum: [full, partial, hash, encrypt]
          description: 脱敏类型
        accessLevel:
          type: string
          enum: [public, internal, private, confidential]
          description: 访问级别
        encryptionKey:
          type: string
          description: 加密密钥标识
    MetaRelation:
      type: object
      required: [relationType, source, target]
      properties:
        id:
          type: string
          description: 关系ID
        relationType:
          type: string
          enum: [one-to-one, one-to-many, many-to-many, reference, aggregation, composition]
          description: 关系类型
        source:
          type: string
          description: 源对象ID
        target:
          type: string
          description: 目标对象ID
        field:
          type: string
          description: 关联字段
        direction:
          type: string
          enum: [inbound, outbound, bidirectional]
          description: 关系方向
        navigable:
          type: boolean
          default: true
          description: 是否可导航
        constraint:
          type: object
          properties:
            cascade:
              type: boolean
              description: 级联操作
            unique:
              type: boolean
              description: 唯一约束
            nullable:
              type: boolean
              description: 可空约束
          description: 关系约束

    # ==================== 请求响应模型 ====================
...
    UpdateMetaObjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [enabled, disabled, locked]
        extensionFields:
          type: object
          additionalProperties: true
    MetaObjectResponse:
      allOf:
        - $ref: '#/components/schemas/MetaObject'
        - type: object
          properties:
            attributes:
              type: array
              items:
                $ref: '#/components/schemas/MetaAttribute'
            relations:
              type: array
              items:
                $ref: '#/components/schemas/MetaRelation'
    MetaObjectDetailResponse:
      allOf:
        - $ref: '#/components/schemas/MetaObjectResponse'
        - type: object
          properties:
            rules:
              type: array
              items:
                $ref: '#/components/schemas/MetaRule'
            extensions:
              type: array
              items:
                $ref: '#/components/schemas/MetaExtension'
...
    MetaRule:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
          description: 规则ID
        name:
          type: string
          description: 规则名称
        type:
          type: string
          enum: [validation, automation, workflow, calculation, permission]
          description: 规则类型
        scope:
          type: string
          description: 作用范围
        expression:
          type: string
          description: 规则表达式
        dsl:
          type: string
          description: DSL表达式
        condition:
          type: string
          description: 触发条件
        action:
          type: string
          description: 执行动作
        message:
          type: string
          description: 提示信息
        enabled:
          type: boolean
          default: true
          description: 是否启用
        priority:
          type: integer
          default: 0
          description: 优先级
        trigger:
          type: string
          enum: [onCreate, onUpdate, onDelete, onSave, onValidate, manual]
          description: 触发时机
        group:
          type: string
          description: 规则组
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MetaExtension:
      type: object
      properties:
        id:
          type: string
        extensionType:
          type: string
          enum: [lineage, workflow, integration, custom]
        handler:
          type: string
          description: 扩展处理器
        config:
          type: object
          additionalProperties: true
          description: 扩展配置
        enabled:
          type: boolean
          default: true

    # ==================== 监控统计模型 ====================
...
    ErrorResponse:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误信息
        details:
          type: string
          description: 详细信息
        code:
          type: string
          description: 错误码
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: 请求路径
        traceId:
          type: string
          description: 追踪ID
    ConflictResponse:
      type: object
      properties:
        message:
          type: string
          default: "资源冲突"
        conflicts:
          type: array
          items:
            type: object
            properties:
              resource:
                type: string
              reason:
                type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "请求参数格式错误"
            code: "INVALID_PARAMETER"
            timestamp: "2024-12-19T10:30:00Z"
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "请求的资源不存在"
            code: "RESOURCE_NOT_FOUND"
            timestamp: "2024-12-19T10:30:00Z"
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictResponse'
          example:
            message: "资源存在依赖关系，无法删除"
            conflicts:
              - resource: "User.orders"
                reason: "存在未完成的订单"
            timestamp: "2024-12-19T10:30:00Z"
...